feat(service): Consolidating service ownership into the cli module
